requirements.txt
crewai
crewai-tools
langchain
langchain-openai
python-dotenv
pydantic
requests
